pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.21",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "tiktoken>=0.12.0",
    "torch>=2.11.0",
    "transformers[torch]>=5.5.0",
    "torchvision>=0.11.0",
    "numpy>=1.21.0",
    "Pillow>=8.0.0",
    "opencv-python>=4.5.0",
    "tqdm>=4.62.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
